+++ remote-wallet/src/lib.rs
@@ -14,3 +14,4 @@ pub mod ledger_error;
 pub mod locator;
 pub mod remote_keypair;
 pub mod remote_wallet;
+pub mod trezor;