+++ remote-wallet/src/remote_keypair.rs
@@ -1,7 +1,7 @@
 use {
     crate::{
-        ledger::get_ledger_from_info,
-        locator::{Locator, Manufacturer},
+        ledger::get_wallet_from_info,
+        locator::Locator,
         remote_wallet::{
             RemoteWallet, RemoteWalletError, RemoteWalletInfo, RemoteWalletManager,
             RemoteWalletType,
@@ -29,6 +29,7 @@ impl RemoteKeypair {
     ) -> Result<Self, RemoteWalletError> {
         let pubkey = match &wallet_type {
             RemoteWalletType::Ledger(wallet) => wallet.get_pubkey(&derivation_path, confirm_key)?,
+            RemoteWalletType::Trezor(wallet) => wallet.get_pubkey(&derivation_path, confirm_key)?,
         };
 
         Ok(Self {
@@ -50,6 +51,9 @@ impl Signer for RemoteKeypair {
             RemoteWalletType::Ledger(wallet) => wallet
                 .sign_message(&self.derivation_path, message)
                 .map_err(|e| e.into()),
+            RemoteWalletType::Trezor(wallet) => wallet
+                .sign_message(&self.derivation_path, message)
+                .map_err(|e| e.into()),
         }
     }
 
@@ -66,16 +70,12 @@ pub fn generate_remote_keypair(
     keypair_name: &str,
 ) -> Result<RemoteKeypair, RemoteWalletError> {
     let remote_wallet_info = RemoteWalletInfo::parse_locator(locator);
-    if remote_wallet_info.manufacturer == Manufacturer::Ledger {
-        let ledger = get_ledger_from_info(remote_wallet_info, keypair_name, wallet_manager)?;
-        let path = format!("{}{}", ledger.pretty_path, derivation_path.get_query());
-        Ok(RemoteKeypair::new(
-            RemoteWalletType::Ledger(ledger),
-            derivation_path,
-            confirm_key,
-            path,
-        )?)
-    } else {
-        Err(RemoteWalletError::DeviceTypeMismatch)
-    }
+    let remote_wallet = get_wallet_from_info(remote_wallet_info, keypair_name, wallet_manager)?;
+    let path = format!("{}{}", remote_wallet.path, derivation_path.get_query());
+    RemoteKeypair::new(
+        remote_wallet.wallet_type,
+        derivation_path,
+        confirm_key,
+        path,
+    )
 }